Add Grappel Prototype
I am experimenting with the use of spring joints for the players grapple ability, I created a simple prototype to test it's limits. At this point it feels very juicy and fun, but I have yet to determine if the joint can control tightly enough for the player to use it accurately and meaningfully in game play. Player movement may need to be tightened and the player size will have to be increased. (or at least player size relative to the camera) A more dynamic camera system is a must. Testing with multiple players is a must.